S2LP Driver: fixed code errors which break compilation
Replaced all calls to `S2LPSpiReadRegisters` with `S2LP_ReadRegister`,
and all calls to `S2LPSpiWriteRegisters` with `S2LP_WriteRegister`.
This fixes the undefined reference to the replaced functions.

Removed duplicate enum definition `ModeExtRef`, which was
originally defined in `s2lp.h`, while leaving a single definition
in `s2lp_general.h`, to align with the intent of `s2lp_general` files.

Defined macro for `s_assert_param` as it was undefined.
Later this macro should be removed (see the additional context of #3).

Included `s2lp_general.h` in `s2lp.h`.
Then removed duplicate functions defined in `s2lp_general.h`:
- `S2LPGeneralGetDevicePN`, replaced with `S2LP_GetDevicePN`
- `S2LPGeneralGetVersion`, replaced with `S2LP_GetVersion`
- `S2LPGeneralSetExtRef`,  replaced with `S2LP_SetExtRef`
- `S2LPGeneralGetExtRef`, replaced with `S2LP_GetExtRef`
- `S2LPRadioSetExternalSmpsMode`, replaced with `S2LP_SetExternalSmpsMode`
- Removed `S2LPRefreshStatus`, since it already
exists (and widely used) as `S2LP_RefreshStatus`.
The replaced functions were moved from `s2lp.c` to `s2lp_general.c`.
Accordingly, moved the function definitions from `s2lp.h` to `s2lp_general.h`
**This way, compatibility with the latest release version is kept**.

Moved the `g_xStatus` declaration to s2lp.h, and removed
the duplicate defitintion of `g_xStatus` from `s2lp_types.c`.
It is more reasonable to define it in `s2lp.h` and `s2lp.c`, since
this is a variable and not a type. This aligns with the definition
of the definition of  `S2LP_RefreshStatus` in the `s2lp.c` file,
as well as `IO_funcs` variable in `s2lp.c`.

Removed the duplicate `S2LPGeneralLibraryVersion()` macro
from `s2lp_general.h`.

This patch was successfully compiled in STM32CubeIde 1.8.0, and
was tested on a custom board with simple P2P app.

Fixes: https://github.com/STMicroelectronics/x-cube-subg2/issues/3
